r6483/sofa-dev : FIX : try to remove some of the warnings. yes some dirty job has to be done sometimes, by the writers should be better (aaahhh! my eyes are blooding seeing some parts of the code)
 //partial sync (11/12 changes)//

+++ modules/sofa/component/collision/AttachBodyPerformer.cpp
@@ -72,7 +72,7 @@ void AttachBodyPerformer<defaulttype::Rigid3fTypes>::start_partial(const BodyPic
 
 
     spring.setInitLength(this->interactor->getMouseRayModel()->getRay(0).direction()*restLength);
-    spring.setSoftStiffnessTranslation(stiffness);
+    spring.setSoftStiffnessTranslation((float)stiffness);
     jointspringforcefield->addSpring(spring);
 
     const core::objectmodel::TagSet &tags=mstateCollision->getTags();
